Overhaul of hardware defines & brightness
To accomidate different hardware configurations, I put in a bunch of defines at the top of LEDLamp. Uncomment the correct defines before compiling for hardware. The checked-in version works for Libby's box.
This update also re-works all the brightness controls so they are consistent between simulations.
Also removed the Firefly simulation because it wasn't as good as the others.